test(rocm): #785 P1 HOLD repair — unlink CTest, exclusive A/B
Shared fixture header, GPU hashes printed/checked, exclusive XOR
classifier, add_executable-only GPU target, pinned A/B env.
Product src/include unchanged vs 52ed5f83. GPU remains HOLD.

FOLLOWING_AGENTS_PROTOCOL

Following-Agents-Protocol: true
AI-Assisted: true
